keyNavigation: split down the engine into smaller parts
