+++ src/plugins/content/accessibility/keyNavigation/keyNavigator.js
@@ -23,16 +23,15 @@
 import $ from 'jquery';
 import _ from 'lodash';
 import keyNavigator from 'ui/keyNavigation/navigator';
-import navigableDomElement from 'ui/keyNavigation/navigableDomElement';
 import navigableGroupElement from 'ui/keyNavigation/navigableGroupElement';
+import {allowedToNavigateFrom} from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/helpers';
+import headerNavigationStrategyFactory from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/strategies/headerNavigation';
+import toolbarNavigationStrategyFactory from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/strategies/toolbarNavigation';
+import navigatorNavigationStrategyFactory from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/strategies/navigatorNavigation';
+import pageNavigationStrategyFactory from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/strategies/pageNavigation';
+import rubricsNavigationStrategyFactory from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/strategies/rubricsNavigation';
+import itemNavigationStrategyFactory from 'taoQtiTest/runner/plugins/content/accessibility/keyNavigation/strategies/itemNavigation';
 import shortcut from 'util/shortcut';
-
-/**
- * When either an element or its parents have this class - navigation from it would be disabled.
- *
- * @type {String}
- */
-const ignoredClass = 'no-key-navigation';
 
 const keysForTypesMap = {
     default: {
@@ -76,104 +75,27 @@ const keysForTypesMap = {
 /**
  * Init the navigation in the toolbar
  *
+ * @param {Object} testRunner
  * @param {Object} config
  * @returns {Array}
  */
-function initToolbarNavigation(config) {
-    const $navigationBar = $('.bottom-action-bar');
-    const $focusables = $navigationBar.find('.action:not(.btn-group):visible, .action.btn-group .li-inner:visible');
-    const elements = navigableDomElement.createFromDoms($focusables);
-    const isNativeNavigation = config.contentNavigatorType === 'native';
-    if (elements.length) {
-        return [
-            keyNavigator({
-                id: 'bottom-toolbar',
-                replace: true,
-                group: $navigationBar,
-                elements: elements,
-                defaultPosition(navigables) {
-                    if (isNativeNavigation) {
-                        return 0;
-                    }
-                    let pos = navigables.length - 1;
-                    // start from the button "Next" or the button "End test"
-                    _.forEach(navigables, (navigable, i) => {
-                        const $element = navigable.getElement();
-                        // find button "Next"
-                        if ($element.data('control') &&
-                            ($element.data('control') === 'move-forward' ||
-                            $element.data('control') === 'move-end')) {
-                            pos = i;
-                        }
-                    });
-                    // else the last button
-                    return pos;
-                }
-            })
-                .on(config.keyNextInGroup, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.next();
-                    }
-                })
-                .on(config.keyPrevInGroup, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.previous();
-                    }
-                })
-                .on('activate', function(cursor) {
-                    cursor.navigable
-                        .getElement()
-                        .click()
-                        .mousedown();
-                })
-        ];
-    }
-    return [];
+function initToolbarNavigation(testRunner, config) {
+    const strategy = toolbarNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
  * Init the navigation in the header block
  *
+ * @param {Object} testRunner
  * @param {Object} config
  * @returns {Array}
  */
-function initHeaderNavigation(config) {
-    //need global selector as currently no way to access delivery frame from test runner
-    const $header = $('header');
-    const $headerElements = $header.find('a:visible');
-    const navigables = navigableDomElement.createFromDoms($headerElements);
-    if (navigables.length) {
-        return [
-            keyNavigator({
-                id: 'header-toolbar',
-                group: $header,
-                elements: navigables,
-                replace: true
-            })
-                .on(config.keyNextInGroup, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.next();
-                    }
-                })
-                .on(config.keyPrevInGroup, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.previous();
-                    }
-                })
-                .on('activate', function(cursor) {
-                    cursor.navigable.getElement().click();
-                })
-        ];
-    }
-    return [];
+function initHeaderNavigation(testRunner, config) {
+    const strategy = headerNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
@@ -184,172 +106,9 @@ function initHeaderNavigation(config) {
  * @returns {Array} the keyNavigator of the main navigation group
  */
 function initNavigatorNavigation(testRunner, config) {
-    const $panel = testRunner.getAreaBroker().getPanelArea();
-    const $navigator = $panel.find('.qti-navigator');
-    const isNativeNavigation = config.contentNavigatorType === 'native';
-    const navigators = [];
-    let filtersNavigator;
-    let itemsNavigator;
-    let $filters, $trees, navigableFilters, navigableTrees;
-
-    //the tag to identify if the item listing has been browsed, to only "smart jump" to active item only on the first visit
-    let itemListingVisited = false;
-    //the position of the filter in memory, to only "smart jump" to active item only on the first visit
-    let filterCursor;
-
-    if ($navigator.length && !$navigator.hasClass('disabled')) {
-        $filters = $navigator.find('.qti-navigator-filters .qti-navigator-filter');
-        navigableFilters = navigableDomElement.createFromDoms($filters);
-        if (navigableFilters.length) {
-            filtersNavigator = keyNavigator({
-                keepState: !isNativeNavigation,
-                id: 'navigator-filters',
-                replace: true,
-                elements: navigableFilters,
-                group: $navigator.find('.qti-navigator-filters')
-            })
-                .on(config.keyNextInFilters, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.next();
-                    }
-                })
-                .on(config.keyPrevInFilters, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.previous();
-                    }
-                })
-                .on('activate', function(cursor) {
-                    cursor.navigable.getElement().click();
-                })
-                .on('focus', function(cursor, origin) {
-                    if (!isNativeNavigation) {
-                        //activate the tab in the navigators
-                        cursor.navigable.getElement().click();
-
-                        //reset the item listing browsed tag whenever the focus on the filter happens after a focus on another element
-                        if ((filterCursor && filterCursor.position !== cursor.position) || origin) {
-                            itemListingVisited = false;
-                        }
-                        //set the filter cursor in memory
-                        filterCursor = cursor;
-                    }
-
-                });
-            if (!isNativeNavigation) {
-                filtersNavigator
-                    .on(config.keyNextInList, function(elem) {
-                        if (!allowedToNavigateFrom(elem)) {
-                            return false;
-                        } else if (itemsNavigator) {
-                            _.defer(function() {
-                                if (itemListingVisited) {
-                                    itemsNavigator.focus().first();
-                                } else {
-                                    itemsNavigator.focus();
-                                }
-                            });
-                        }
-                    })
-                    .on(config.keyPrevInList, function(elem) {
-                        if (!allowedToNavigateFrom(elem)) {
-                            return false;
-                        } else if (itemsNavigator) {
-                            _.defer(function() {
-                                itemsNavigator.last();
-                            });
-                        }
-                    });
-            }
-            navigators.push(filtersNavigator);
-        }
-
-        const $navigatorTree = $panel.find('.qti-navigator-tree');
-        $trees = $navigator.find('.qti-navigator-tree .qti-navigator-item:not(.unseen) .qti-navigator-label');
-        navigableTrees = navigableDomElement.createFromDoms($trees);
-        if (navigableTrees.length) {
-            //instantiate a key navigator but do not add it to the returned list of navigators as this is not supposed to be reached with tab key
-            itemsNavigator = keyNavigator({
-                id: 'navigator-items',
-                replace: true,
-                elements: navigableTrees,
-                group: $navigatorTree,
-                defaultPosition(navigables) {
-                    let pos = 0;
-                    if (filterCursor && filterCursor.navigable.getElement().data('mode') !== 'flagged') {
-                        _.forEach(navigables, function(navigable, i) {
-                            const $parent = navigable.getElement().parent('.qti-navigator-item');
-                            //find the first active and visible item
-                            if ($parent.hasClass('active') && $parent.is(':visible')) {
-                                pos = i;
-                                return false;
-                            }
-                        });
-                    }
-                    return pos;
-                }
-            })
-                .on(config.keyNextInList, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.next();
-                    }
-                })
-                .on(config.keyPrevInList, function(elem) {
-                    if (!allowedToNavigateFrom(elem)) {
-                        return false;
-                    } else {
-                        this.previous();
-                    }
-                })
-                .on('activate', function(cursor) {
-                    cursor.navigable.getElement().click();
-                })
-                .on('lowerbound upperbound', function() {
-                    if (!isNativeNavigation && filtersNavigator) {
-                        filtersNavigator.focus();
-                    }
-                })
-                .on('focus', function(cursor) {
-                    itemListingVisited = true;
-                    cursor.navigable
-                        .getElement()
-                        .parent()
-                        .addClass('key-navigation-highlight');
-                })
-                .on('blur', function(cursor) {
-                    cursor.navigable
-                        .getElement()
-                        .parent()
-                        .removeClass('key-navigation-highlight');
-                });
-            if (!isNativeNavigation) {
-                itemsNavigator
-                    .on(config.keyNextInFilters, function(elem) {
-                        if (!allowedToNavigateFrom(elem)) {
-                            return false;
-                        } else if (filtersNavigator) {
-                            filtersNavigator.focus().next();
-                        }
-                    })
-                    .on(config.keyPrevInFilters, function(elem) {
-                        if (!allowedToNavigateFrom(elem)) {
-                            return false;
-                        } else if (filtersNavigator) {
-                            filtersNavigator.focus().previous();
-                        }
-                    });
-            }
-            if (isNativeNavigation) {
-                navigators.push(itemsNavigator);
-            }
-        }
-    }
-    return navigators;
+    const strategy = navigatorNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
@@ -362,33 +121,9 @@ function initNavigatorNavigation(testRunner, config) {
  * @returns {Array} of keyNavigator ids
  */
 function initDefaultContentNavigation(testRunner, config) {
-    let itemNavigators = [];
-    const $content = testRunner.getAreaBroker().getContentArea();
-
-    //the item focusable body elements are considered scrollable
-    $content.find('.key-navigation-focusable').addClass('key-navigation-scrollable');
-    $content
-        .find('.key-navigation-focusable,.qti-interaction')
-        .filter(function() {
-            //filter out interaction as it will be managed separately
-            return !$(this).parents('.qti-interaction').length;
-        })
-        .each(function() {
-            const $itemElement = $(this);
-            if ($itemElement.hasClass('qti-interaction')) {
-                itemNavigators = _.union(itemNavigators, initInteractionNavigation($itemElement, testRunner, config));
-            } else {
-                itemNavigators.push(
-                    keyNavigator({
-                        elements: navigableDomElement.createFromDoms($itemElement),
-                        group: $itemElement,
-                        propagateTab: false
-                    })
-                );
-            }
-        });
-
-    return itemNavigators;
+    const strategy = itemNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
@@ -399,171 +134,24 @@ function initDefaultContentNavigation(testRunner, config) {
  * @param {Object} testRunner
  * @returns {Array} of keyNavigator ids
  */
-function initAllContentButtonsNavigation(testRunner) {
-    const navigableElements = [];
-    const $content = testRunner.getAreaBroker().getContentArea();
-    const $qtiIteractionsNodeList = $content.find('.key-navigation-focusable,.qti-interaction').filter(function() {
-        //filter out interaction as it will be managed separately
-        return !$(this).parents('.qti-interaction').length;
-    });
-    const $qtiChoiceNodesList = $qtiIteractionsNodeList.find('.qti-choice');
-
-    //the item focusable body elements are considered scrollable
-    $content.find('.key-navigation-focusable').addClass('key-navigation-scrollable');
-
-    $qtiChoiceNodesList.each(function() {
-        const $itemElement = $(this);
-        const keyNavigatorItem = keyNavigator({
-            elements: navigableDomElement.createFromDoms($itemElement),
-            group: $itemElement,
-            propagateTab: false
-        });
-
-        keyNavigatorItem.on('activate', function(cursor) {
-            const $elt = cursor.navigable.getElement();
-            //jQuery <= 1.9.0 the checkbox values are set
-            //after the click event if triggerred with jQuery
-            if ($elt.is(':checkbox')) {
-                $elt.each(function() {
-                    this.click();
-                });
-            } else {
-                $elt.click();
-            }
-        });
-
-        navigableElements.push(keyNavigatorItem);
-    });
-
-    return navigableElements;
-}
-
-/**
- * Init interaction key navigation from the interaction navigator
- *
- * @param {jQuery} $interaction - the interaction container
- * @param {Object} testRunner
- * @param {Object} config
- * @returns {Array} array of navigators created from interaction container
- */
-function initInteractionNavigation($interaction, testRunner, config) {
-    let $inputs;
-    let interactionNavigables;
-    const interactionNavigators = [];
-
-    //add navigable elements from prompt
-    $interaction.find('.key-navigation-focusable').each(function() {
-        const $nav = $(this);
-        if (!$nav.closest('.qti-choice').length) {
-            interactionNavigators.push(
-                keyNavigator({
-                    elements: navigableDomElement.createFromDoms($nav),
-                    group: $nav,
-                    propagateTab: false
-                })
-            );
-        }
-    });
-
-    //reset interaction custom key navigation to override the behaviour with the new one
-    $interaction.off('.keyNavigation');
-
-    //search for inputs that represent the interaction focusable choices
-    $inputs = $interaction.is(':input') ? $interaction : $interaction.find(':input');
-    interactionNavigables = navigableDomElement.createFromDoms($inputs);
-
-    if (interactionNavigables.length) {
-        const keyNavigatorItem = keyNavigator({
-            elements: interactionNavigables,
-            group: $interaction,
-            loop: false
-        });
-
-        keyNavigatorItem
-            .on(config.keyNextInGroup, function(elem) {
-                if (!allowedToNavigateFrom(elem)) {
-                    return false;
-                } else {
-                    this.next();
-                }
-            })
-            .on(config.keyPrevInGroup, function(elem) {
-                if (!allowedToNavigateFrom(elem)) {
-                    return false;
-                } else {
-                    this.previous();
-                }
-            })
-            .on('activate', function(cursor) {
-                const $elt = cursor.navigable.getElement();
-
-                //jQuery <= 1.9.0 the checkbox values are set
-                //after the click event if triggerred with jQuery
-                if ($elt.is(':checkbox')) {
-                    $elt.each(function() {
-                        this.click();
-                    });
-                } else {
-                    $elt.click();
-                }
-            })
-            .on('focus', function(cursor) {
-                const $qtiChoice = cursor.navigable.getElement().closest('.qti-choice');
-                $qtiChoice.addClass('key-navigation-highlight');
-                showElementsContent($qtiChoice, testRunner.getAreaBroker().getContentArea());
-            })
-            .on('blur', function(cursor) {
-                cursor.navigable
-                    .getElement()
-                    .closest('.qti-choice')
-                    .removeClass('key-navigation-highlight');
-            });
-
-        interactionNavigators.push(keyNavigatorItem);
-    }
-
-    return interactionNavigators;
-}
-
-/**
- * Scrolling to the top of the required element
- * @param {jQuery} $el
- * @param {jQuery} $visibleContainer
- */
-function showElementsContent($el, $visibleContainer) {
-    const $wrapper = $visibleContainer.closest('.content-wrapper');
-    if ($wrapper.length && $el.length) {
-        $wrapper.scrollTop($el.offset().top + $wrapper.scrollTop() - $wrapper.offset().top);
-    }
+function initAllContentButtonsNavigation(testRunner, config) {
+    const strategy = itemNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
  * Init the navigation of test rubric blocks
  * It returns an array of keyNavigator ids as the content is dynamically determined
  *
+ * @param {Object} testRunner
+ * @param {Object} config
  * @returns {Array} of keyNavigator ids
  */
-function initRubricNavigation() {
-    let $itemElements;
-    const rubricNavigators = [];
-    const $rubricArea = $('#qti-rubrics');
-
-    $itemElements = $rubricArea.find('.qti-rubricBlock');
-    $itemElements.each(function() {
-        const $itemElement = $(this);
-        const id = `rubric_element_navigation_group_${rubricNavigators.length}`;
-
-        rubricNavigators.push(
-            keyNavigator({
-                id: id,
-                elements: navigableDomElement.createFromDoms($itemElement),
-                group: $itemElement,
-                replace: true
-            })
-        );
-    });
-
-    return rubricNavigators;
+function initRubricNavigation(testRunner, config) {
+    const strategy = rubricsNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
@@ -571,32 +159,13 @@ function initRubricNavigation() {
  * make this element scrollable
  *
  * @param {Object} testRunner
+ * @param {Object} config
  * @returns {Array} of keyNavigator ids
  */
-function initDefaultItemNavigation(testRunner) {
-    const itemNavigators = [];
-    const $container = $(testRunner.getAreaBroker().getContainer());
-    const $wrapper = $container.find('.content-wrapper');
-    let keyNavigatorItem;
-    const navigables = navigableDomElement.createFromDoms($wrapper);
-
-    $wrapper.addClass('key-navigation-scrollable');
-    if (navigables.length) {
-        const itemId = testRunner.getCurrentItem().id;
-        keyNavigatorItem = keyNavigator({
-            id: `item-content-wrapper_${itemId}`,
-            group: $wrapper,
-            elements: navigables,
-            propagateTab: false, // inner item navigators will send tab to this element
-            replace: true,
-        });
-
-        itemNavigators.push(
-            keyNavigatorItem
-        );
-    }
-
-    return itemNavigators;
+function initDefaultItemNavigation(testRunner, config) {
+    const strategy = pageNavigationStrategyFactory(testRunner, config);
+    strategy.init();
+    return strategy.getNavigators();
 }
 
 /**
@@ -617,31 +186,31 @@ function initTestRunnerNavigation(testRunner, pluginConfig) {
     switch (config.contentNavigatorType) {
         case 'linear':
             navigators = _.union(
-                initRubricNavigation(),
-                initAllContentButtonsNavigation(testRunner),
-                initToolbarNavigation(config),
-                initHeaderNavigation(config),
+                initRubricNavigation(testRunner, config),
+                initAllContentButtonsNavigation(testRunner, config),
+                initToolbarNavigation(testRunner, config),
+                initHeaderNavigation(testRunner, config),
                 initNavigatorNavigation(testRunner, config),
-                initDefaultItemNavigation(testRunner)
+                initDefaultItemNavigation(testRunner, config)
             );
             break;
         case 'native':
             navigators = _.union(
-                initHeaderNavigation(config),
+                initHeaderNavigation(testRunner, config),
                 initNavigatorNavigation(testRunner, config),
-                initRubricNavigation(),
+                initRubricNavigation(testRunner, config),
                 initDefaultContentNavigation(testRunner, config),
-                initToolbarNavigation(config)
+                initToolbarNavigation(testRunner, config)
             );
             break;
         default:
             navigators = _.union(
-                initRubricNavigation(),
+                initRubricNavigation(testRunner, config),
                 initDefaultContentNavigation(testRunner, config),
-                initToolbarNavigation(config),
-                initHeaderNavigation(config),
+                initToolbarNavigation(testRunner, config),
+                initHeaderNavigation(testRunner, config),
                 initNavigatorNavigation(testRunner, config),
-                initDefaultItemNavigation(testRunner)
+                initDefaultItemNavigation(testRunner, config)
             );
             break;
     }
@@ -708,22 +277,6 @@ function initTestRunnerNavigation(testRunner, pluginConfig) {
     }
 
     return keyNavigatorItem;
-}
-
-/**
- * Checks whether element is navigable from
- *
- * @param {HTMLElement} element
- * @returns {boolean}
- */
-function allowedToNavigateFrom(element) {
-    const $element = $(element);
-
-    if ($element.hasClass(ignoredClass) || $element.parents(`.${ignoredClass}`).length > 0) {
-        return false;
-    }
-
-    return true;
 }
 
 /**
